requirements.txt
pandas
beautifulsoup4
requests
tabulate
python-dotenv
